requirements.txt
scapy
streamlit
requests
python-dotenv
scikit-learn
numpy
pandas
